CLEANUP: reorganize for the cryptocurrency
