+++ scripts/average-work.js
@@ -1,4 +1,4 @@
-const Blockchain = require('./blockchain');
+const Blockchain = require('../blockchain');
 
 
 const blockchain = new Blockchain();
